Cleanup in the prelude.
Many ugly implementation details were moved to an 'implementation/'
subdirectory. Also some layout and documentation fixes, and a few other
tidbits.
